# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Collaborative Notion-like Teaching Reference API
  version: 1.0.0
  description: |
    RESTful API for collaborative document editing with DDD, Clean Architecture, and CRDT-based real-time sync.

    ## Authentication
    All endpoints require Bearer token authentication except /auth/login and /auth/register.

    ## Rate Limiting
    Not implemented in MVP.

    ## Versioning
    API version in path (future: /v1/orgs).
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.notion-clone.example.com
    description: Production
security:
  - bearerAuth: []
paths:
...
  /orgs/{id}/pages:
    post:
      summary: Create page in organization
      operationId: createPage
      tags: [Documents]
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePageRequest'
      responses:
        '201':
          description: Page created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'

    get:
      summary: List pages in organization
      operationId: listPages
      tags: [Documents]
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: List of pages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PageDto'
  /pages/{id}:
    get:
      summary: Get page details
      operationId: getPage
      tags: [Documents]
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Page details with blocks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageDetailDto'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    CreatePageRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
    PageDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        orgId:
          type: string
          format: uuid
        title:
          type: string
        createdBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
    PageDetailDto:
      allOf:
        - $ref: '#/components/schemas/PageDto'
        - type: object
          properties:
            blocks:
              type: array
              items:
                $ref: '#/components/schemas/BlockDto'
...
    BlockDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        pageId:
          type: string
          format: uuid
        parentBlockId:
          type: string
          format: uuid
          nullable: true
        sortKey:
          type: number
          format: double
        type:
          type: string
          enum: [paragraph, heading, todo, file]
        json:
          type: object
...
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...